requirements.txt
minemeld-core>=0.9.31b2
libtaxii>=1.1.111
requests
pytz>=2017.2